chore: rename lexical keys & folder names to camelCase
