+++ packages/richtext-lexical/src/field/features/debug/testRecorder/plugin/index.scss
