+++ packages/richtext-lexical/src/field/features/debug/treeView/plugin/index.scss
