+++ packages/richtext-lexical/src/field/features/horizontalRule/component/index.tsx
