+++ packages/richtext-lexical/src/field/features/horizontalRule/feature.client.tsx
