+++ packages/richtext-lexical/src/field/features/horizontalRule/plugin/index.scss
